CliconServer/requirements.txt
django
pillow
django-rest-framework
django-cors-headers
